Disable comment hiding on topsport.bg
Website uses javascript request to fetch data from partner.googleadservices.com, and if it fails it blocks reading and writing of comments below the articles. Google DFP configuration is stored in a separate file (dfp.js), which is blocked by most filter, so no ads are served despite adding this rule.