Politiken.dk update
On www.politiken.dk they now have a randomly placed modal div with a random CSS class. The javascript that generates it is bundled with the rest of their scripts.

So basically, this is a mini-whitelist of their main divs rather than trying to identify and block the adblock blocker.